feat(OBT-6): Create a separate component for the grid calendar
