+++ src/app/features/goals/goals.component.html
@@ -1,1 +1,1 @@
-<p>goals works!</p>
+<app-grid-calendar />
