+++ src/app/features/goals/grid-calendar/grid-calendar.component.html
@@ -1,0 +1,1 @@
+<p>grid-calendar works!</p>
