+++ src/app/features/goals/grid-calendar/grid-calendar.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { GridCalendarComponent } from './grid-calendar.component';
+
+@NgModule({
+  declarations: [GridCalendarComponent],
+  imports: [CommonModule],
+  exports: [GridCalendarComponent],
+})
+export class GridCalendarModule {}
